Moved GCPMetricDescriptor to own file
